backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-text-splitters
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
pymupdf
pdfplumber
tiktoken
groq
python-multipart
scikit-learn
joblib
numpy
